+++ components/src/lib/Modal/ModalInput.tsx
@@ -1,0 +1,5 @@
+import { Input } from "./styles";
+
+export const ModalInput = ({ placeholder, onChange }: { placeholder: string; onChange?: () => void }) => {
+	return <Input placeholder={placeholder} onChange={onChange} />;
+};
